+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 0.2.1b0
 -------
+* New app rascil_advise to advise on imaging parameters for a MeasurementSet
 * Implement using requirements*.in files for dependency management.
   Add files: requirements.in, requirements-docs.in, requirements-test.in
 * Makefile updated with requirements management commands: